scripts/requirements.txt
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
mouseinfo==0.1.3
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyautogui==0.9.54
pydantic==2.10.6
pydantic-core==2.27.2
pygetwindow==0.0.9
pymsgbox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
pyrect==0.2.0
pyscreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.1.0
python3-xlib==0.15
pytweening==1.2.0
requests==2.32.3
rsa==4.9
tqdm==4.67.1
typing-extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
